aider: Import RowWinChecker and ColWinChecker in IWinChecker.js.
# Aider chat conversation:

USER: ReferenceError: RowWinChecker is not defined
    at IWinChecker.CreateWinChecker (file:///C:/Users/user/Documents/Coding/JavaScriptSlotMachineOO/IWinChecker.js:7:21)
    at Board.checkWins (file:///C:/Users/user/Documents/Coding/JavaScriptSlotMachineOO/board.js:30:39)
    at AccountMgr.calcPnL (file:///C:/Users/user/Documents/Coding/JavaScriptSlotMachineOO/account_mgr.js:27:33)
    at SlotMachine.playRound (file:///C:/Users/user/Documents/Coding/JavaScriptSlotMachineOO/slot_machine.js:27:42)
    at Game.start (file:///C:/Users/user/Documents/Coding/JavaScriptSlotMachineOO/app.js:12:25)
    at file:///C:/Users/user/Documents/Coding/JavaScriptSlotMachineOO/app.js:25:6
    at ModuleJob.run (node:internal/modules/esm/module_job:194:25)

Node.js v18.18.0

ASSISTANT: The error message indicates that `RowWinChecker` is not defined in `IWinChecker.js`. This is because we are trying to create an instance of `RowWinChecker` and `ColWinChecker` in `IWinChecker.js` without importing them.

To fix this, we need to import `RowWinChecker` and `ColWinChecker` in `IWinChecker.js`.

Here is the *SEARCH/REPLACE* block:

IWinChecker.js
```javascript
<<<<<<< SEARCH
class IWinChecker {
=======
import RowWinChecker from "./RowWinChecker.js";
import ColWinChecker from "./ColWinChecker.js";

class IWinChecker {
>>>>>>> REPLACE
```

Please apply this change to your code.